삭제 기능 UX 개선

+++ vue-til/src/components/posts/PostListItem.vue
@@ -22,8 +22,11 @@ export default {
   },
   methods: {
     async deleteItem() {
-      await deletePost(this.postItem._id);
-      console.log('deleted');
+      if (confirm('You want to delete it?')) {
+        await deletePost(this.postItem._id);
+        this.$emit('refresh');
+      }
+      // console.log('deleted');
     },
   },
 };
